py/pyproject.toml
[project]
name = "skkuverse-crawler"
version = "1.0.0"
description = "SKKU 관련 데이터 크롤링 + 콘텐츠 정제 서비스"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "nh3>=0.2",
    "motor>=3.6",
    "apscheduler>=3.10,<4",
    "structlog>=24.0",
    "click>=8.1",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.7",
    "mypy>=1.13",
]

[project.scripts]
skkuverse-crawler = "skkuverse_crawler.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skkuverse_crawler"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
